feat: 폰트 설정 분리 및 로컬에서도 관리
